Make "size" members of type usize
The C API exposes blazesym_sym_info type with a size member of type u64.
We use usize for various sizes so let's do the same here. I went ahead
and converted a few more internal types as well. I preserved the
blazesym_sym_info.file_offset as u64 because that's how Rust generally
represents file offset (e.g.,
https://doc.rust-lang.org/std/io/enum.SeekFrom.html).

+++ src/c_api.rs
@@ -535,7 +535,7 @@ pub unsafe extern "C" fn blazesym_result_free(results: *const blazesym_result) {
 pub struct blazesym_sym_info {
     pub name: *const c_char,
     pub address: Addr,
-    pub size: u64,
+    pub size: usize,
     pub file_offset: u64,
     pub obj_file_name: *const c_char,
     pub sym_type: blazesym_sym_type,
